hw-6/requirements.txt
greenlet==3.0.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.26
typing_extensions==4.9.0
